### 2017-05-08 #254 @arokem
@cMadan : would you like to review this for JOSS?
