### 2017-05-12 #259 @whedon
[REVIEW]: Neuropsydia.py: A Python Module for Creating Experiments, Tasks and Questionnaires
**Submitting author:** @DominiqueMakowski (<a href="http://orcid.org/0000-0001-5375-9967">Dominique Makowski</a>)
**Repository:** <a href="https://github.com/neuropsychology/Neuropsydia.py/" target ="_blank">https://github.com/neuropsychology/Neuropsydia.py/</a>
**Version:** 1.0.4
