### 2017-05-12 #259 @cMadan
@DominiqueMakowski - I just did a quick first pass at going through the JOSS reviewer questions, but not yet testing the functionality of the software itself. There are a few things I noticed that need to be addressed, and I will continue and test the code itself within a few days.

Version - You list the software as being version 1.0.4 in the repo, but haven't minted a 'release' in Github to go with this. If someone wanted to specifically run version 1.0.4 in the future, this wouldn't be possible. Also, JOSS will later require you to archive the accepted version of the software in an archival repository (such as Zenodo or figshare), but that isn't necessary yet.

Documentation - the API documentation link in the README is listed as http://neuropsydia.readthedocs.io/en/master/documentation.html, but this appears to only be a skeleton document. In the paper, the link is http://neuropsydia.readthedocs.io/en/latest/documentation.html, which does have content. These should both be linked to the same complete document.

References - DOIs are currently missing.
