### 2017-06-11 #259 @arokem
@DominiqueMakowski : have you had a chance to address the comments that @cMadan made?
