### 2017-06-19 #259 @DominiqueMakowski
@arokem I plan to address them by the end of the week. Sorry for the delay!
