### 2017-06-23 f13475f
Added DOIs
