+++ paper/paper.bib
@@ -1,6 +1,7 @@
 @article{krause2014expyriment,
   title={Expyriment: A Python library for cognitive and neuroscientific experiments},
   author={Krause, Florian and Lindemann, Oliver},
+  doi = {0.3758/s13428-013-0390-6},
   journal={Behavior Research Methods},
   volume={46},
   number={2},
@@ -12,6 +13,7 @@
 @article{mathot2012opensesame,
   title={OpenSesame: An open-source, graphical experiment builder for the social sciences},
   author={Math{\^o}t, Sebastiaan and Schreij, Daniel and Theeuwes, Jan},
+  doi = {10.3758/s13428-011-0168-7},
   journal={Behavior research methods},
   volume={44},
   number={2},
@@ -24,6 +26,7 @@
 @article{peirce2007psychopy,
   title={PsychoPy—psychophysics software in Python},
   author={Peirce, Jonathan W},
+  doi = {10.1016/j.jneumeth.2006.11.017},
   journal={Journal of neuroscience methods},
   volume={162},
   number={1},
