### 2017-06-23 #259 @DominiqueMakowski
Dear @arokem, sorry for the long answer, and thanks a ***lot*** to @cMadan for his first comments.

**Version - You list the software as being version 1.0.4 in the repo, but haven't minted a 'release' in Github to go with this. If someone wanted to specifically run version 1.0.4 in the future, this wouldn't be possible. Also, JOSS will later require you to archive the accepted version of the software in an archival repository (such as Zenodo or figshare), but that isn't necessary yet.**

That has been done. We created a 1.0.4 release on github.

**Documentation - the API documentation link in the README is listed as http://neuropsydia.readthedocs.io/en/master/documentation.html, but this appears to only be a skeleton document. In the paper, the link is http://neuropsydia.readthedocs.io/en/latest/documentation.html, which does have content. These should both be linked to the same complete document.**

We replaced the wrong links with the correct one.

**References - DOIs are currently missing.**

This has been modified [here](https://github.com/neuropsychology/Neuropsydia.py/blob/master/paper/paper.bib) accordingly.

Again, thanks a lot and sorry for the delay.
