### 2017-06-24 d9c9b95
Added choices_style parameter
Enables support for chinese characters in choice
