+++ neuropsydia.egg-info/SOURCES.txt
@@ -1,0 +1,65 @@
+LICENSE
+MANIFEST.in
+README.md
+README.rst
+setup.cfg
+setup.py
+neuropsydia/__init__.py
+neuropsydia/ask.py
+neuropsydia/choice.py
+neuropsydia/core.py
+neuropsydia/image.py
+neuropsydia/io.py
+neuropsydia/meta.py
+neuropsydia/miscellaneous.py
+neuropsydia/path.py
+neuropsydia/procedures.py
+neuropsydia/scale.py
+neuropsydia/screen.py
+neuropsydia/start.py
+neuropsydia/write.py
+neuropsydia.egg-info/PKG-INFO
+neuropsydia.egg-info/SOURCES.txt
+neuropsydia.egg-info/dependency_links.txt
+neuropsydia.egg-info/requires.txt
+neuropsydia.egg-info/top_level.txt
+neuropsydia/files/__init__.py
+neuropsydia/files/binary/__init__.py
+neuropsydia/files/binary/absorption_desk.png
+neuropsydia/files/binary/absorption_man.png
+neuropsydia/files/binary/cursor_absorption.png
+neuropsydia/files/binary/cursor_absorption_validated.png
+neuropsydia/files/binary/cursor_blue.png
+neuropsydia/files/binary/cursor_blue_validated.png
+neuropsydia/files/binary/cursor_green.png
+neuropsydia/files/binary/cursor_green_validated.png
+neuropsydia/files/binary/cursor_orange.png
+neuropsydia/files/binary/cursor_orange_validated.png
+neuropsydia/files/binary/cursor_purple.png
+neuropsydia/files/binary/cursor_purple_validated.png
+neuropsydia/files/binary/cursor_red.png
+neuropsydia/files/binary/cursor_red_validated.png
+neuropsydia/files/binary/cursor_teal.png
+neuropsydia/files/binary/cursor_teal_validated.png
+neuropsydia/files/binary/cursor_yellow.png
+neuropsydia/files/binary/cursor_yellow_validated.png
+neuropsydia/files/binary/scale_point.png
+neuropsydia/files/font/LiberationMono-Bold.ttf
+neuropsydia/files/font/LiberationMono-Regular.ttf
+neuropsydia/files/font/RobotoBlack.ttf
+neuropsydia/files/font/RobotoBold.ttf
+neuropsydia/files/font/RobotoLight.ttf
+neuropsydia/files/font/RobotoRegular.ttf
+neuropsydia/files/font/Timeless-Bold.ttf
+neuropsydia/files/font/Timeless.ttf
+neuropsydia/files/font/__init__.py
+neuropsydia/files/logo/N.png
+neuropsydia/files/logo/Neuropsydia_HEAD_white.png
+neuropsydia/files/logo/Neuropsydia_PSY_blue.png
+neuropsydia/files/logo/Neuropsydia_TEXT_white.png
+neuropsydia/files/logo/Python.png
+neuropsydia/files/logo/__init__.py
+neuropsydia/files/logo/icon.png
+neuropsydia/files/logo/neuropsydia_banner.png
+neuropsydia/tests/__init__.py
+neuropsydia/tests/test_color.py
