### 2017-07-23 #259 @arokem
Thanks - ping me here when things are resolved, with pointers to the relevant issues. Thanks again!
