+++ neuropsydia/data.py
@@ -1,0 +1,136 @@
+# -*- coding: utf-8 -*-
+import os
+
+import pandas as pd
+import numpy as np
+
+
+
+
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+def read_data(filename, extension="", participant_id="", path="", localization="US", print_warning=True):
+    """
+    Load the datafile into a pandas' dataframe. Made for the neuropsydia package.
+
+    Parameters
+    ----------
+    NA
+
+    Returns
+    ----------
+    NA
+
+    Example
+    ----------
+    NA
+
+    Authors
+    ----------
+    Dominique Makowski
+
+    Dependencies
+    ----------
+    - pandas
+    """
+    # Find a corresponding file
+    file = filename
+    if os.path.isfile(file) is False:
+        file = path + filename + extension
+    if os.path.isfile(file) is False:
+        file = path + filename + ".xlsx"
+    if os.path.isfile(file) is False:
+        file = path + filename + ".csv"
+    if os.path.isfile(file) is False:
+        file = path + participant_id + filename + extension
+    if os.path.isfile(file) is False:
+        if ".csv" in file:
+            file = path + "/csv/" + participant_id + "_" + filename + extension
+        elif ".xlsx" in file:
+            file = path + "/excel/" + participant_id + "_" + filename + extension
+        else:
+            extension = ".xlsx"
+    if os.path.isfile(file) is False:
+        if print_warning is True:
+            print("NeuroKit Error: read_data(): file's path " + file + " not found!")
+
+    if localization == "FR" or localization == "FRA" or localization == "French" or localization == "France":
+        sep = ";"
+        decimal = ","
+    else:
+        sep = ","
+        decimal = "."
+
+    if ".csv" in file:
+        try:
+            df = pd.read_csv(file, sep=sep, decimal=decimal, encoding="utf-8")
+        except UnicodeDecodeError:
+            df = pd.read_csv(file, sep=sep, decimal=decimal, encoding="cp1125")
+    elif ".xls" in file or ".xlsx" in file:
+        df = pd.read_excel(file, encoding="utf-8")
+    else:
+        if print_warning is True:
+            print("NeuroKit Error: read_data(): wrong extension of the datafile.")
+    return(df)
+
+
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+def save_data(df, filename="data", extension="all", participant_id="", path="", localization="US", index=False, print_warning=True, index_label=None):
+    """
+    Save the datafile into a pandas' dataframe.
+
+    Parameters
+    ----------
+    NA
+
+    Returns
+    ----------
+    NA
+
+    Example
+    ----------
+    NA
+
+    Authors
+    ----------
+    Dominique Makowski
+
+    Dependencies
+    ----------
+    - pandas
+    """
+    if localization == "FR" or localization == "FRA" or localization == "French" or localization == "France":
+        sep = ";"
+        decimal = ","
+    else:
+        sep = ","
+        decimal = "."
+
+    if extension == "all":
+        extension = [".csv", ".xlsx"]
+
+    for ext in list(extension):
+        if ext == ".csv":
+            if os.path.exists(path + "/csv/") is False:
+                os.makedirs(path + "/csv/")
[…]
