### 2017-11-10 #259 @cMadan
@arokem I was able to test the package fine on a Windows 10 PC and it worked great. Also, as mentioned last time another person has had it work on a Mac, so we can leave that debugging for another time. The package does what's supposed to do, is well documented, and the paper seems to be in order. I recommend the submission be accepted at JOSS.

@DominiqueMakowski, sorry for the delay in testing it further!
