### 2017-11-14 #259 @arokem
Great! @DominiqueMakowski :  could you please create an archive (e.g., on zenodo) of the final version of the software and share the DOI here?
