### 2017-11-14 #259 @DominiqueMakowski
@cMadan Thanks a lot for all your suggestions and comments.

@arokem , the DOI is the following: 10.5281/zenodo.1048777

[![DOI](https://zenodo.org/badge/70266940.svg)](https://zenodo.org/badge/latestdoi/70266940)

Thank you for your work!
