### 2017-11-14 #259 @arokem
@arfon : I believe this paper can now be published.
